[IMP] sale_timesheet: profitability report
In the Project Overview, the Profitability report takes into account
Timesheet costs, but doesn't include information on Revenue.
Because of that, the user is missing crucial information to determine
the profitability of a given project.

This commit aims at fixing this issue by sending the user to the
'Project Costs and Revenues' report of the concerned project when
clicking the 'Profitability' button within the Project Overview page.

Task-1963932

closes odoo/odoo#32554

Signed-off-by: Jérome Maes (jem) <jem@openerp.com>