[FIX] base_gengo: private key password field
Before this commit, the gengo private password was displayed in
plaintext in edit mode. This is due to text fields being rendered as
`textarea` in edit mode, which does not support the "password" type.

This commit fixes the issue by turning this field to `Char`, so that
it is rendered as `input` with type "password" in edit mode.

The loss of multi-line from `textarea` is not a big deal, because
gengo passwords are not that long.

Task-ID 1892058

closes odoo/odoo#32725

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>