[FIX] account: multi-currency payment matching
- Set the company currency in USD, activate multi-curency
- Create an invoice in EUR
- Create a payment in EUR for less than the invoice
- Open the payment matching => all amounts are in USD
- Create a write-off, reconcile

The write-off amount is incorrectly converted.

The write-off amount is sent in USD, but considered in EUR in the
server since the currency considered is the one of the line. However,
the currency in the widget is computed based on the account currency:

https://github.com/odoo/odoo/blob/ab03cbae25d24c511e3eab583d697d6fee7737c3/addons/account/models/account_move.py#L654

Note that this is the opposite of commit 3cfd4aebfdc7deeecf421c293b3e1,
introduced in v9. Since the widget was refactored in v11, many changes
have been introduced, and we assume that this change is now justified.

opw-1907332

closes odoo/odoo#32723

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>