[IMP] base,product,website_sale: move image fallback mixin -> variant
Backport of 3ef1e32b5a1d6e785a71bc4bb32a75f31486e86f

Only the variant is using this code, so it was a bad idea to make the mixin
more complicated than it had to be.

Update the image tests to make sure this is working correctly. The test were
only testing the size of the images, now they also test the actual content.

Part of task-1949729

Original PR: odoo/odoo#32253

closes odoo/odoo#32814

Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>