[IMP] web_gantt: allow creation on gantt to be a wizard
Previous behaviour was to use a formview inside a dialog to create
object on gantt view. This is still the default but now it is possible
to call a wizard, which brings more control over the creation dialog.

In order to do so, it was necessary to allow for  new field on the gantt
arch, whih will hold the xmlid of the wizard's action

related task #1929079