[FIX] digest: using greaterthan and lessthan builtins
Before this commit, when you send a digest template to the user,
the user would receive a blank mail(without any content).
And while rendering the template we were getting the following error:
jinja2.exceptions.TemplateSyntaxError: unexpected char '&' at 4039

This was due to the changes made in commit:
https://github.com/odoo/odoo/commit/5d16f5ec84980e80a2f969bf8f5f7535cf2cb80d#diff-705145eec648d5ef61b7b4060ae96b9dR8
In this instead of using CDATA, there was a change of the field type to 'html' and thus there were issues of parsing for jinja.

Now after this commit, we will be using builtins of jinja:
 - 'greaterthan' instead of '>'
 - 'lessthan' instead of '<'

Task-1918364